+++ Paltarumi.Acopio.Balanza.Dto/Chancado/Mapa/UpdateMapaDto.cs
@@ -5,6 +5,6 @@ namespace Paltarumi.Acopio.Balanza.Dto.Chancado.Mapa
     public class UpdateMapaDto
     {
         public int IdLoteChancado { get; set; }
-        public float Tmh { get; set; }
+        public decimal Tmh { get; set; }
     }
 }
